Håndter halvt grunnbeløp på samme måte for alle brev.

+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/behandling/avslag/Avslag.kt
@@ -10,5 +10,5 @@ data class Avslag(
     val harEktefelle: Boolean,
     val beregning: Beregning?,
     val formuegrunnlag: Formuegrunnlag?,
-    val halvtGrunnbeløpPerÅr: Double,
+    val halvtGrunnbeløpPerÅr: Int,
 )
